:recycle: Migrate Stop constants to TravelSegments.
