+++ web/src/api/stops-api.js
@@ -1,7 +1,7 @@
 import http from "@/api/http-client"
 
 // TODO: fetch accommodation types from backend,
-// until then, keep in sync with src/api/services/estimates/bulk-generate.ts
+// until then, keep in sync with api/src/models/travel-segment.ts
 export const ACCOMMODATION_TYPES = Object.freeze({
   HOTEL: "Hotel",
   PRIVATE: "Private",
@@ -9,6 +9,7 @@ export const ACCOMMODATION_TYPES = Object.freeze({
 })
 
 // TODO: load from back-end
+// until then, keep in sync with api/src/models/travel-segment.ts
 export const TRAVEL_METHODS = Object.freeze({
   AIRCRAFT: "Aircraft",
   POOL_VEHICLE: "Pool Vehicle",
@@ -18,6 +19,12 @@ export const TRAVEL_METHODS = Object.freeze({
   OTHER: "Other:",
 })
 
+/*
+DEPRECATED: Whenever you use this model, try and figure out how to migrate
+the functionality to the TravelSegment model instead.
+It was too large a project to migrate to the TravelSegment model all at once,
+so we're doing it piecemeal.
+*/
 export const stopsApi = {
   TRAVEL_METHODS,
   ACCOMMODATION_TYPES,
